Fix bool/int case in arithmetic interpreter
Really our interpreter should interpret expressions of different types
than just Int (so it could return an Int *or* a Bool), so a comparison
on Ints should return a Bool.  But we can figure out what to do about
that later; right now we are only handling core integer arithmetic.